// src/Umbraco.Community.YouTubePicker/Client/src/api/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts
export type YouTubeSearchResult = {
  kind: string;
  etag: string;
  nextPageToken: string;
  previousPageToken: string;
  regionCode: string;
  pageInfo: PageInfo | null;
  items: Array<(YouTubeItem)> | null;
  status: number;
  statusText: string;
};

export type YouTubeItem = {
  kind: string;
  etag: string;
  id: Id,
  snippet: Snippet
};

export type PageInfo = {
  totalResults: number,
  resultsPerPage: number
};

export type Id = {
  kind: string;
  videoId: string;
  playlistId: string;
}

export type Snippet = {
  publishedAt: Date;
  channelId: string;
  title: string;
  description: string;
  thumbnails: ThumbnailType;
  channelTitle: string;
  liveBroadcastContent: string;
  publishTime: Date;
}

export type ThumbnailType = {
  default: ThumbnailDetail;
  medium: ThumbnailDetail;
  high: ThumbnailDetail;
}

export type ThumbnailDetail = {
  url: string;
  width: number;
  height: number;
}